+++ SensorFusion/src/application/SystemTasksManager/SystemTasksManager.h
@@ -23,9 +23,10 @@ private:
 
 	TaskParams m_taskParam;
 	static inline std::unique_ptr<business_logic::ImageCapturer3D> m_image3DCapturer;
-
-	std::shared_ptr<business_logic::ClockSyncronization::SharedClockSlaveManager> m_globalClkMng;
+	static inline std::shared_ptr<business_logic::ClockSyncronization::SharedClockSlaveManager> m_globalClkMng;
 	std::shared_ptr<business_logic::Communication::CommunicationManager> m_commMng;
+	static inline uint64_t m_lastCaptureTimestampStart = 0;
+	static inline uint64_t m_lastCaptureTimestampEnd = 0;
 
 	static inline std::shared_ptr<business_logic::Osal::TaskHandler> m_clockSyncTaskHandler;
 	static inline std::shared_ptr<business_logic::Osal::TaskHandler> m_image3dCapturerTaskHandler;