+++ SensorFusion/src/business_logic/DataSerializer/Image3DSnapshot.h
@@ -30,7 +30,7 @@ class Image3DSnapshot : public ISerializableMessage
 {
 public:
 	Image3DSnapshot() = default;
-	Image3DSnapshot(uint8_t msgId, uint8_t m_msgIndex, const std::shared_ptr<std::array<LidarPoint, IMAGE3D_SIZE>> image3d, uint16_t image3dSize, uint32_t timestamp);
+	Image3DSnapshot(uint8_t msgId, uint8_t m_msgIndex, const std::shared_ptr<std::array<LidarPoint, IMAGE3D_SIZE>> image3d, uint16_t image3dSize, uint64_t timestamp, uint64_t captureDeltaTime);
 	~Image3DSnapshot() = default;
 
     void serialize(std::vector<uint8_t>& serializedData) const override;
@@ -40,6 +40,7 @@ public:
     	return (m_msgId       == other.m_msgId &&
     			m_msgIndex    == other.m_msgIndex &&
 				m_timestamp   == other.m_timestamp &&
+				m_captureDeltaTime   == other.m_captureDeltaTime &&
 				m_image3dSize == other.m_image3dSize &&
 				m_image3d     == other.m_image3d);
 	}
@@ -49,14 +50,14 @@ public:
     uint8_t  m_msgIndex;
     std::shared_ptr<std::array<LidarPoint, IMAGE3D_SIZE>> m_image3d;
     uint16_t m_image3dSize;
-    uint32_t m_timestamp;
-
+    uint64_t m_timestamp;
+    uint64_t m_captureDeltaTime;
     friend void to_json(nlohmann::json& j, const Image3DSnapshot& image)
     {
 		try
 		{
 			j = nlohmann::json{{"msgId", image.m_msgId}, {"imgSize", image.m_image3dSize},
-							   {"m_image3d", *image.m_image3d}, {"timestamp", image.m_timestamp}};
+							   {"m_image3d", *image.m_image3d}, {"timestamp", image.m_timestamp}, {"captureDeltaTime", image.m_captureDeltaTime}};
 		}
 		catch (const std::exception& e)
 		{
@@ -69,7 +70,9 @@ public:
     	image.m_msgId = j.at("msgId").get<uint8_t>();
     	image.m_image3dSize = j.at("imgSize").get<size_t>();
         *image.m_image3d = j.template get<std::array<LidarPoint, IMAGE3D_SIZE>>();
-    	image.m_timestamp = j.at("timestamp").get<uint32_t>();
+    	image.m_timestamp = j.at("timestamp").get<uint64_t>();
+    	image.m_captureDeltaTime = j.at("captureDeltaTime").get<uint64_t>();
+
     }
 };
 }